+++ @navikt/core/react/src/tabs/Tab.tsx
@@ -2,8 +2,8 @@ import * as RadixTabs from "@radix-ui/react-tabs";
 import cl from "clsx";
 import React, { forwardRef, useContext } from "react";
 import { BodyShort } from "../typography";
-import { TabsContext } from "./Tabs";
 import { OverridableComponent } from "../util";
+import { TabsContext } from "./context";
 
 export interface TabProps
   extends Omit<React.HTMLAttributes<HTMLButtonElement>, "children"> {
