+++ @navikt/core/react/src/tabs/index.ts
@@ -1,4 +1,4 @@
-export { default as Tabs, type TabsProps } from "./Tabs";
 export { type TabProps } from "./Tab";
 export { type TabListProps } from "./TabList";
 export { type TabPanelProps } from "./TabPanel";
+export { default as Tabs, type TabsProps } from "./Tabs";
